First version of new memory manager done.

+++ src/arc/mm/map.c
@@ -48,7 +48,7 @@ static const char *mm_map_type_desc(int type)
 
 static void mm_map_add(int type, uintptr_t addr_start, uintptr_t addr_end)
 {
-  for (int id = 0; id < map.count; id++)
+  for (size_t id = 0; id < map.count; id++)
   {
     mm_map_entry_t *entry = &map.entries[id];
     if (entry->type == MULTIBOOT_MMAP_UNUSED)
@@ -74,14 +74,14 @@ static void mm_map_sanitize(void)
 {
   /* do as many passes as required to fully simplify the map */
 next_pass:
-  for (int id = 0; id < map.count; id++)
+  for (size_t id = 0; id < map.count; id++)
   {
     /* get entry, skip if it is unused */
     mm_map_entry_t entry = map.entries[id];
     if (entry.type == MULTIBOOT_MMAP_UNUSED)
       continue;
 
-    for (int inner_id = 0; inner_id < map.count; inner_id++)
+    for (size_t inner_id = 0; inner_id < map.count; inner_id++)
     {
       /* don't compare an entry with itself */
       if (id == inner_id)
@@ -143,7 +143,7 @@ next_pass:
 
   map.count = 0;
 
-  for (int id = 0; id < tmp_map.count; id++)
+  for (size_t id = 0; id < tmp_map.count; id++)
   {
     mm_map_entry_t *entry = &tmp_map.entries[id];
     if (entry->type != MULTIBOOT_MMAP_UNUSED)
@@ -161,7 +161,7 @@ next_pass:
     bool swap = false;
 
     /* iterate through the list swapping adjacent pairs if required */
-    for (int id = 0; id < map.count - 1; id++)
+    for (size_t id = 0; id < map.count - 1; id++)
     {
       if (map.entries[id].addr_start > map.entries[id + 1].addr_start)
       {
@@ -230,7 +230,7 @@ mm_map_t *mm_map_init(multiboot_t *multiboot)
   mm_map_sanitize();
 
   /* print the final map */
-  for (int id = 0; id < map.count; id++)
+  for (size_t id = 0; id < map.count; id++)
   {
     uintptr_t start = map.entries[id].addr_start;
     uintptr_t end = map.entries[id].addr_end;
